+++ car-bridge/Cargo.toml
@@ -9,8 +9,10 @@ async-channel = "1.7"
 async-stream = "0.3"
 async-trait = { workspace = true }
 chariott-common = { workspace = true }
+chariott-proto = { workspace = true }
 futures = { workspace = true }
 paho-mqtt = "0.11"
+prost = { workspace = true }
 tokio = { workspace = true, features = ["rt-multi-thread"] }
 tokio-stream = { workspace = true }
 tracing = { workspace = true }
