+++ car-bridge/src/lib.rs
@@ -1,4 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
+pub mod chariott;
 pub mod messaging;
